Verifica;óes extras

+++ instalar_minikube.sh
@@ -15,21 +15,25 @@
 
 set -e
 
-echo -e '==== Instalando o Minikube ===='
-curl -LO https://storage.googleapis.com/minikube/releases/latest/minikube-linux-amd64
-sudo install minikube-linux-amd64 /usr/bin/minikube && rm minikube-linux-amd64
+MINIKUBE=$(minikube status 2>&1)
 
-echo -e '\n==== Iniciando o Minikube ===='
-minikube start
+elif [[ $MINIKUBE =~ "command not found" ]]
+  echo -e '==== Instalando o Minikube\n'
+  curl -LO https://storage.googleapis.com/minikube/releases/latest/minikube-linux-amd64
+  sudo install minikube-linux-amd64 /usr/bin/minikube && rm minikube-linux-amd64
 
-echo -e '\n==== Habilitando os addons necessários no Minikube ===='
+if [[ $MINIKUBE =~ "To start a cluster" ]]; then
+  echo -e '==== Iniciando o Minikube\n'
+  minikube start
+
+echo -e '\n==== Habilitando os addons necessários no Minikube'
 minikube addons enable ingress
 minikube addons enable metrics-server
 
-echo -e '\n==== Incluindo os endereços no arquivo de hosts ===='
+echo -e '\n==== Incluindo os endereços no arquivo de hosts'
 sudo bash -c "echo -e '\n# Ambiente do Projeto ADA/Caixa' >> /etc/hosts"
 sudo bash -c "echo '$(minikube ip) minio.projeto.ada' >> /etc/hosts"
 sudo bash -c "echo '$(minikube ip) rabbitmq.projeto.ada' >> /etc/hosts"
 sudo bash -c "echo '$(minikube ip) redis.projeto.ada' >> /etc/hosts"
 
-echo -e '\n==== Instalação finalizada ===='
+echo -e '\n==== Instalação finalizada'
